修复私有的ctype_val基类模板函数还能类外访问的编译器问题

+++ src/engine/z3_target_call/AMD64/amd64CCall.h
@@ -4,12 +4,12 @@ extern "C" {
 #include "engine/basic_var.hpp"
 #include "z3_target_defs.h"
 
-rsbool z3_amd64g_calculate_rflags_cf(ULong cc_op, const rsval<uint64_t>& cc_dep1_formal, const rsval<uint64_t>& cc_dep2_formal, const rsval<uint64_t>& cc_ndep_formal);
-rsbool z3_amd64g_calculate_rflags_pf(ULong cc_op, const rsval<uint64_t>& cc_dep1_formal, const rsval<uint64_t>& cc_dep2_formal, const rsval<uint64_t>& cc_ndep_formal);
-rsbool z3_amd64g_calculate_rflags_af(ULong cc_op, const rsval<uint64_t>& cc_dep1_formal, const rsval<uint64_t>& cc_dep2_formal, const rsval<uint64_t>& cc_ndep_formal);
-rsbool z3_amd64g_calculate_rflags_zf(ULong cc_op, const rsval<uint64_t>& cc_dep1_formal, const rsval<uint64_t>& cc_dep2_formal, const rsval<uint64_t>& cc_ndep_formal);
-rsbool z3_amd64g_calculate_rflags_sf(ULong cc_op, const rsval<uint64_t>& cc_dep1_formal, const rsval<uint64_t>& cc_dep2_formal, const rsval<uint64_t>& cc_ndep_formal);
-rsbool z3_amd64g_calculate_rflags_of(ULong cc_op, const rsval<uint64_t>& cc_dep1_formal, const rsval<uint64_t>& cc_dep2_formal, const rsval<uint64_t>& cc_ndep_formal);
+rsbool z3_amd64g_calculate_rflags_cf(int cc_op, const rsval<uint64_t>& cc_dep1_formal, const rsval<uint64_t>& cc_dep2_formal, const rsval<uint64_t>& cc_ndep_formal);
+rsbool z3_amd64g_calculate_rflags_pf(int cc_op, const rsval<uint64_t>& cc_dep1_formal, const rsval<uint64_t>& cc_dep2_formal, const rsval<uint64_t>& cc_ndep_formal);
+rsbool z3_amd64g_calculate_rflags_af(int cc_op, const rsval<uint64_t>& cc_dep1_formal, const rsval<uint64_t>& cc_dep2_formal, const rsval<uint64_t>& cc_ndep_formal);
+rsbool z3_amd64g_calculate_rflags_zf(int cc_op, const rsval<uint64_t>& cc_dep1_formal, const rsval<uint64_t>& cc_dep2_formal, const rsval<uint64_t>& cc_ndep_formal);
+rsbool z3_amd64g_calculate_rflags_sf(int cc_op, const rsval<uint64_t>& cc_dep1_formal, const rsval<uint64_t>& cc_dep2_formal, const rsval<uint64_t>& cc_ndep_formal);
+rsbool z3_amd64g_calculate_rflags_of(int cc_op, const rsval<uint64_t>& cc_dep1_formal, const rsval<uint64_t>& cc_dep2_formal, const rsval<uint64_t>& cc_ndep_formal);
 
  rsval<uint64_t> z3_amd64g_calculate_condition(const rsval<uint64_t>/*AMD64Condcode*/& cond, const rsval<uint64_t>& cc_op, const rsval<uint64_t>& cc_dep1, const rsval<uint64_t>& cc_dep2, const rsval<uint64_t>& cc_ndep);
  rsval<uint64_t> z3_amd64g_calculate_rflags_c(const rsval<uint64_t>& cc_op, const rsval<uint64_t>& cc_dep1, const rsval<uint64_t>& cc_dep2, const rsval<uint64_t>& cc_ndep);
